+++ src/controllers/adminControllers/loginController.ts
@@ -1,26 +1,26 @@
 import { Request, Response } from "express";
-import AdminSchema from "../model/adminSchema";
-import { generateToken } from "../utils/jwt";
-import { comparePassword } from "../utils/bcrypt";
+import AdminSchema from "../../models/adminModels/adminSchema";
+import { generateToken } from "../../utils/jwt";
+import { comparePassword } from "../../utils/bcrypt";
 
 const login = async (req: Request, res: Response) => {
   interface loginBody {
-    Email: string;
-    Password: string;
+    email: string;
+    password: string;
   }
 
   try {
-    const { Email, Password }: loginBody = req.body;
+    const { email, password }: loginBody = req.body;
 
-    if (!Email || !Password) {
+    if (!email || !password) {
       res.status(401).json({
         success: false,
-        message: "Email or PhoneNumber and Password are required",
+        message: "Email or PhoneNumber and password are required",
       });
       return;
     }
 
-    const admin = await AdminSchema.findOne({ Email });
+    const admin = await AdminSchema.findOne({ email });
 
     if (!admin) {
       res.status(402).json({
@@ -30,12 +30,12 @@ const login = async (req: Request, res: Response) => {
       return;
     }
 
-    const validatedAdmin = await comparePassword(Password, admin?.Password);
+    const validatedAdmin = await comparePassword(password, admin?.password);
 
     if (!validatedAdmin) {
       res.status(403).json({
         success: false,
-        message: "Invalid Password",
+        message: "Invalid password",
       });
       return;
     }
