+++ src/controllers/adminControllers/otpController.ts
@@ -1,20 +1,20 @@
 import { Request, Response } from "express";
-import AdminSchema from "../model/adminSchema";
-import pendingAdmins from "../utils/pendingAdmin";
-import { verifyOTP } from "../utils/otp";
+import AdminSchema from "../../models/adminModels/adminSchema";
+import pendingAdmins from "../../utils/pendingAdmin";
+import { verifyOTP } from "../../utils/otp";
 
 export const verifySignUpOTP = async (req: Request, res: Response) => {
   try {
-    const { Email, otp }: { Email: string; otp: string } = req.body;
+    const { email, otp }: { email: string; otp: string } = req.body;
 
-    if (!Email || !otp) {
+    if (!email || !otp) {
       res
         .status(400)
         .json({ success: false, message: "Email and OTP are required." });
       return;
     }
 
-    const isValidOTP = verifyOTP(Email, otp);
+    const isValidOTP = verifyOTP(email, otp);
     if (!isValidOTP) {
       res
         .status(403)
@@ -22,7 +22,7 @@ export const verifySignUpOTP = async (req: Request, res: Response) => {
       return;
     }
 
-    const adminData = pendingAdmins[Email];
+    const adminData = pendingAdmins[email];
     if (!adminData) {
       res
         .status(404)
@@ -31,16 +31,16 @@ export const verifySignUpOTP = async (req: Request, res: Response) => {
     }
 
     const newAdmin = new AdminSchema({
-      FirstName: adminData.FirstName ,
-      LastName: adminData.LastName ,
-      Email: adminData.Email,
-      PhoneNumber: adminData.PhoneNumber,
-      Password: adminData.Password,
+      FirstName: adminData.firstName ,
+      LastName: adminData.lastName ,
+      email: adminData.email,
+      PhoneNumber: adminData.phoneNumber,
+      Password: adminData.password,
       isVerified: true,
     });
 
     await newAdmin.save();
-    delete pendingAdmins[Email];
+    delete pendingAdmins[email];
 
     res.status(200).json({
       success: true,
